+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.2] - 2023-08-01
+### Added
+- [@mjkuranda](https://github.com/mjkuranda): Maximal entity count in map up to 32 entities.
+
+### Fixed
+- [@mjkuranda](https://github.com/mjkuranda): Rendering entities in `RetroRenderer`.
+
 ## [0.15.1] - 2023-07-29
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): Space char in `InputMenuOption`.